Move presentational components to view dir
